[Refactor] Rename private disjunctive and agent task graph plotting modules
